[feat] question 기능구현

+++ server/overflow/src/main/java/com/preproject/overflow/question/service/QuestionService.java
@@ -1,7 +1,7 @@
 package com.preproject.overflow.question.service;
 
-import com.preproject.overflow.question.exception.ExceptionCode;
-import com.preproject.overflow.question.exception.BusinessLogicException;
+import com.preproject.overflow.exception.ExceptionCode;
+import com.preproject.overflow.exception.BusinessLogicException;
 import com.preproject.overflow.question.entity.Question;
 import com.preproject.overflow.question.entity.QuestionTag;
 import com.preproject.overflow.question.repository.QuestionRepository;
@@ -52,13 +52,17 @@ public class QuestionService {
     }
 
     // 게시글 내용 수정
-    public Question updateQuestion(Question question) {
+    public Question updateQuestion(Question question, Member loggedInMember) {
         Question findQuestion = questionRepository.findByQuestionId(question.getQuestionId());
 
-        Optional.ofNullable(question.getText())
-                .ifPresent(text -> findQuestion.setText(text));
+        if (findQuestion.getMember().getMemberId().equals(loggedInMember.getMemberId())) {
+            Optional.ofNullable(question.getText())
+                    .ifPresent(text -> findQuestion.setText(text));
 
-        return questionRepository.save(findQuestion);
+            return questionRepository.save(findQuestion);
+        } else {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PERMISSION_DENIED);
+        }
     }
 
     public Question voteQuestion(Question question, Boolean vote) {
@@ -125,12 +129,15 @@ public class QuestionService {
         return questionPage;
     }
 
-    public void deleteQuestion(Long questionId) {
+    public void deleteQuestion(Long questionId, Member loggedInMember) {
         Question question = findVerifyQuestion(questionId);
+
+        if (!question.getMember().getMemberId().equals(loggedInMember.getMemberId())) {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PERMISSION_DENIED);
+        }
         deleteTag(question);
         //downQuestionCount(question);
         questionRepository.delete(question);
-
     }
 
     public void createTag(Question question) {
